% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_analysis_clusterProfiler.R
\name{batch_GO}
\alias{batch_GO}
\title{Run Batch GO Analysis}
\usage{
batch_GO(gene_lists, orgdb, ...)
}
\arguments{
\item{gene_lists}{Named list of gene vectors}

\item{orgdb}{OrgDb object}

\item{...}{Additional parameters passed to pGO}
}
\value{
List of GO enrichment results
}
\description{
Run Batch GO Analysis
}
