% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_pdf_dim.R
\name{calculate_pdf_dimensions}
\alias{calculate_pdf_dimensions}
\title{Calculate PDF Dimensions for Multiple Plots}
\usage{
calculate_pdf_dimensions(n_plots, ncol = 2, square_size = 8, margin = 0.15)
}
\arguments{
\item{n_plots}{Integer. The total number of plots to be arranged.}

\item{ncol}{Integer. Number of columns in the plot grid layout. Default is 2.}

\item{square_size}{Numeric. The size (in inches) of each plot square. Default is 8.}

\item{margin}{Numeric. The margin factor as a proportion of square size. Default is 0.15.}
}
\value{
A list containing:
\itemize{
\item width: The calculated width of the PDF document in inches
\item height: The calculated height of the PDF document in inches
}
}
\description{
This function calculates the appropriate width and height for a PDF document
that will contain multiple plots arranged in a grid layout.
}
\examples{
# Calculate dimensions for 5 plots
dims <- calculate_pdf_dimensions(5)

# Calculate dimensions with custom settings
dims <- calculate_pdf_dimensions(
  n_plots = 6,
  ncol = 3,
  square_size = 6,
  margin = 0.2
)

}
