% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGeneEffLengthAlt.R
\name{compareGeneLengthMethods}
\alias{compareGeneLengthMethods}
\title{Compare Gene Length Calculation Methods}
\usage{
compareGeneLengthMethods(gtf_file)
}
\arguments{
\item{gtf_file}{Path to GTF file}
}
\value{
Data frame comparing results from both methods
}
\description{
Compare Gene Length Calculation Methods
}
