% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSCdim.R
\name{get_plot_data}
\alias{get_plot_data}
\title{Get plot data for dimensional reduction visualization}
\usage{
get_plot_data(seurat_obj, reduction, group_by, split_by = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{reduction}{Dimensional reduction to use}

\item{group_by}{Group by variable}

\item{split_by}{Split by variable}
}
\value{
A list containing plot data and dimension names
}
\description{
Get plot data for dimensional reduction visualization
}
\keyword{internal}
