% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_minmax.R
\name{mf_minmax}
\alias{mf_minmax}
\title{Normalize Values to \link{0, 1} Range}
\usage{
mf_minmax(x)
}
\arguments{
\item{x}{A numeric vector or matrix to be normalized.}
}
\value{
If \code{x} is a vector, returns a normalized vector.
If \code{x} is a matrix, returns a matrix with each row normalized.
}
\description{
This function normalizes the input values to a range between 0 and 1 using min-max normalization.
It can handle both vectors and matrices. For matrices, normalization is applied row-wise.
}
\details{
The function uses the formula: (x - min(x)) / (max(x) - min(x))
If all values in a vector or a row of a matrix are the same (i.e., min = max),
a warning is issued and the original values are returned for that vector or row.
}
\examples{
# Vector example
vec <- c(1, 2, 6, 9, 4, 4, 7.2)
mf_minmax(vec)

# Matrix example
mat <- matrix(c(1, 2, 3, 4, 5, 6), nrow = 2, byrow = TRUE)
mf_minmax(mat)

}
