% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGeneEffLength.R
\name{validateGeneLengths}
\alias{validateGeneLengths}
\title{Validate Gene Length Results}
\usage{
validateGeneLengths(gene_lengths)
}
\arguments{
\item{gene_lengths}{Data frame output from calcGeneEL}
}
\value{
Logical indicating if the results pass basic validation
}
\description{
Validate Gene Length Results
}
