% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGeneEffLength.R
\name{calcGeneEL}
\alias{calcGeneEL}
\title{Calculate Effective Gene Lengths from GTF File}
\usage{
calcGeneEL(gtf_file, cores = NULL)
}
\arguments{
\item{gtf_file}{Character string specifying the path to the GTF file}

\item{cores}{Numeric value indicating the number of CPU cores to use (default is half of available cores)}
}
\value{
A data frame with two columns:
\item{geneid}{Gene identifiers}
\item{efflen}{Effective lengths of genes}
}
\description{
This function calculates the effective length of genes by summing up the
non-overlapping exon lengths for each gene from a GTF annotation file.
}
\examples{
\dontrun{
gene_lengths <- calcGeneEL("path/to/your/annotation.gtf.gz")
head(gene_lengths)
}

}
