% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSCdim.R
\name{get_theme_settings}
\alias{get_theme_settings}
\title{Get theme settings for single cell plots}
\usage{
get_theme_settings()
}
\value{
A ggplot2 theme object
}
\description{
Get theme settings for single cell plots
}
\keyword{internal}
